aquel_bridge/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
edge-tts==6.1.9
aiofiles==23.2.1
Pillow==10.2.0
httpx==0.27.0
python-dotenv==1.0.1
